requirements.txt
annotated-types==0.7.0
anyio==4.11.0
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.1
ecdsa==0.19.1
fastapi==0.118.0
idna==3.10
passlib==1.7.4
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pypdfium2==4.30.0
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
